Move hasTag function into tags dir and reuse in listTargetInstances
