--- src/asgs/hasTag.js
@@ -1,3 +1,0 @@
-module.exports = function hasTag(asg, target) {
-  return asg.Tags.some(tag => tag.Key === target);
-};
