+++ src/tags/hasTag.js
@@ -1,0 +1,3 @@
+module.exports = function hasTag(tags, target) {
+  return tags.some(tag => tag.Key === target);
+};
